+++ invisible_backend/src/perpetual/liquidations/liquidation_engine.rs
@@ -103,21 +103,7 @@ impl LiquidationSwap {
             println!("leftover_collateral: {}", leftover_collateral);
             println!("is_partial_liquidation: {}", is_partial_liquidation);
 
-            let new_idx = state_tree.lock().first_zero_idx() as u32;
-
-            //         liquidation_order: &LiquidationOrder,
-            // liquidated_size: u64,
-            // liquidator_fee: u64,
-            // market_price: u64,
-            // current_funding_index: u32,
-            // new_idx: u32,
-
-            // liquidated_size: 48920440
-            // liquidator_fee: 7.674.637 404 000
-            // leftover_collateral: -7674592164155
-            // is_partial_liquidation: false
-            // liquidation swap executed successfully
-            // Position liquidated successfully!!!!!!!!!
+            let new_idx = state_tree.lock().first_zero_idx();
 
             let new_position = open_new_position_after_liquidation(
                 &self.liquidation_order,
